requirements.txt
langchain==0.3.19
langchain-community==0.3.18
python-dotenv==1.0.1
jinja2==3.1.3
python-dateutil==2.8.2
openai==1.65.3
flask==3.0.3
waitress==2.1.2